Групповой проект, тесты 995/995 (#16)
* Реализовано удаление фильмов и пользователей по id (#6)

* Реализовано удаление фильмов и пользователей по id

* Реализовано удаление фильмов и пользователей по id, + ON DELETE CASCADE

* Реализация вывода общих фильмов (#8)

* Реализация вывода общих вильмов

* Правки после ревью.


* Добавлена ветка add-director, тесты 492/492 (#9)

* feature: 43 tests down

* bux-fixing: 7 tests down

* feature: add-directors 492/492 tests

* refactor: внесены правки после первого внутреннего ревью

* refactor: скорректирован конструктор в Film

* refactor: фикс тестов

* refactor: удалены лишние пробелы в Film

* Реализация поиска

feature: реализация поиска

* Реализация отзывов (#11)

* add-review 417/53

* add-review 470/0 нужно оптимизировать валидацию пользователей и фильмов при запросе отзыва

* refactor: минорные правки

* Update ReviewDbStorage.java

Исправления в дизлайках и сортировке при получении всех обзоров из базы

* Убрана сортировка стримом при получении отзывов

* refactor: удалён пробел


* Реализация вывода самых популярных фильмов по жанру и годам

feature : Реализация вывода самых популярных фильмов по жанру и годам


* Реализация рекомендаций

Реализация рекомендаций


* Функциональность «Лента событий»

* Реализация поиска #2

Дополнения к ветке add-search: уменьшено количество запросов в бд


* Fix: диалект изменен на PostgreSQL

Fix: диалект изменен на PostgreSQL

* Update README.md (#19)

Update README.md

* Правки после первого ревью

refactor: правки после первого ревью

---------

+++ src/main/java/ru/yandex/practicum/filmorate/controllers/GenreController.java
@@ -1,7 +1,6 @@
 package ru.yandex.practicum.filmorate.controllers;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.model.FilmGenre;
 import ru.yandex.practicum.filmorate.service.GenreService;
@@ -21,7 +20,7 @@ public class GenreController {
     }
 
     @GetMapping("/{id}")
-    public FilmGenre getById(@PathVariable int id) {
+    public FilmGenre getById(@PathVariable Long id) {
         return genreService.getById(id);
     }
