+++ src/main/java/ru/yandex/practicum/filmorate/storage/genre/GenreStorage.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 public interface GenreStorage {
     List<FilmGenre> getAll();
 
-    Optional<FilmGenre> getById(int id);
+    Optional<FilmGenre> getById(Long id);
 
     FilmGenre create(FilmGenre genre);
